Rename connection config file to db
